Pir sensor bug
I operated it for about 30 minutes in which i was in the room, left the room, close to sensor, far from sensor, however all result is motion detected. This would either be that all heat is detected as human activity while it might just be the lamp or charger, or it is super sensitive to everything which would need to be changed and solved.